separate tooltip components
