+++ packages/ui/src/lib/InfoButton.svelte
@@ -1,92 +1,161 @@
 <script lang="ts">
-	import TooltipWrapper from './shared/TooltipWrapper.svelte';
-	import { pxToRem } from './utils/pxToRem';
+	// import { fly } from 'svelte/transition';
+	import { flyScale } from '$lib/utils/transitions';
 	import type { Snippet } from 'svelte';
 
-	const GAP = 2;
-	const EXTENDED_HOVER_HEIGHT = 4;
-
 	interface Props {
+		title?: string;
+		size?: 'small' | 'medium';
 		children: Snippet;
 	}
 
-	const { children }: Props = $props();
-	const gap = GAP - EXTENDED_HOVER_HEIGHT;
+	const { title, size = 'medium', children }: Props = $props();
 
-	let forceShow: boolean = $state(false);
+	let show = $state(false);
+	let timeoutId: undefined | ReturnType<typeof setTimeout> = $state();
 
 	function handleMouseEnter() {
-		forceShow = true;
+		timeoutId = setTimeout(() => {
+			show = true;
+		}, 700);
 	}
 
 	function handleMouseLeave() {
-		forceShow = false;
+		clearTimeout(timeoutId);
+		show = false;
 	}
 </script>
 
-{#snippet tooltip()}
-	<div
-		class="info-tooltip-container"
-		onmouseenter={handleMouseEnter}
-		onmouseleave={handleMouseLeave}
-		style:padding-top={pxToRem(EXTENDED_HOVER_HEIGHT)}
-		role="tooltip"
-	>
-		<div class="arrow-container">
-			<div class="arrow"></div>
-		</div>
-		<div class="info-tooltip">
-			{@render children()}
-		</div>
-	</div>
-{/snippet}
+<div class="wrapper" role="tooltip" onmouseenter={handleMouseEnter} onmouseleave={handleMouseLeave}>
+	<div class="info-button {size}"></div>
 
-<TooltipWrapper {tooltip} {gap} position="bottom" align="center" {forceShow}>
-	<div>
-		<svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
-			<path
-				fill-rule="evenodd"
-				clip-rule="evenodd"
-				d="M2.5 8C2.5 4.96243 4.96243 2.5 8 2.5C11.0376 2.5 13.5 4.96243 13.5 8C13.5 11.0376 11.0376 13.5 8 13.5C4.96243 13.5 2.5 11.0376 2.5 8ZM1 8C1 4.13401 4.13401 1 8 1C11.866 1 15 4.13401 15 8C15 11.866 11.866 15 8 15C4.13401 15 1 11.866 1 8ZM9 8.5C9 7.94772 8.55229 7.5 8 7.5C7.44772 7.5 7 7.94772 7 8.5L7 10.5C7 11.0523 7.44772 11.5 8 11.5C8.55228 11.5 9 11.0523 9 10.5V8.5ZM9 5.5C9 4.94772 8.55229 4.5 8 4.5C7.44772 4.5 7 4.94772 7 5.5C7 6.05228 7.44772 6.5 8 6.5C8.55229 6.5 9 6.05228 9 5.5Z"
-				fill="#867E79"
-			/>
-		</svg>
-	</div>
-</TooltipWrapper>
+	{#if show}
+		<div class="tooltip-container" transition:flyScale>
+			<div class="tooltip-arrow"></div>
+
+			<div class="tooltip-card">
+				{#if title}
+					<h3 class="text-13 text-semibold tooltip-title">{title}</h3>
+				{/if}
+				<p class="text-12 text-body tooltip-description">
+					{@render children()}
+				</p>
+			</div>
+		</div>
+	{/if}
+</div>
 
 <style lang="postcss">
-	.info-tooltip-container {
-		display: flex;
-		flex-direction: column;
-		align-items: center;
+	.wrapper {
+		position: relative;
+		display: inline-flex;
 	}
 
-	.arrow-container {
+	.info-button {
+		position: relative;
+		flex-shrink: 0;
+		color: var(--clr-text-2);
+		border-radius: 16px;
+		box-shadow: inset 0 0 0 1.5px var(--clr-text-2);
+
+		&::before,
+		&::after {
+			content: '';
+			position: absolute;
+			left: 50%;
+			transform: translateX(-50%);
+			background-color: var(--clr-text-2);
+			border-radius: 2px;
+		}
+	}
+
+	.info-button.medium {
+		width: 16px;
+		height: 16px;
+
+		&::before {
+			top: 4px;
+			width: 2px;
+			height: 2px;
+		}
+
+		&::after {
+			top: 7px;
+			width: 2px;
+			height: 5px;
+		}
+	}
+
+	.info-button.small {
+		width: 12px;
+		height: 12px;
+
+		&::before {
+			top: 3px;
+			width: 2px;
+			height: 2px;
+		}
+
+		&::after {
+			top: 6px;
+			width: 2px;
+			height: 3px;
+		}
+	}
+
+	.tooltip-container {
+		z-index: var(--z-blocker);
+		position: absolute;
+		top: 100%;
+		left: 50%;
+		transform: translateX(-50%);
+		display: flex;
+		flex-direction: column;
+	}
+
+	.tooltip-card {
+		display: flex;
+		flex-direction: column;
+		gap: 4px;
+		background-color: var(--clr-bg-1);
+		border: 1px solid var(--clr-border-2);
+		border-radius: var(--radius-m);
+		padding: 12px;
+		width: max-content;
+		max-width: 260px;
+		box-shadow: var(--fx-shadow-m);
+	}
+
+	.tooltip-title {
+		color: var(--clr-text-1);
+	}
+
+	.tooltip-description {
+		color: var(--clr-scale-ntrl-40);
+	}
+
+	.tooltip-arrow {
 		position: relative;
 		top: 1px;
+		margin: 0 auto;
 		width: 100%;
 		height: 10px;
 		display: flex;
 		justify-content: center;
 		overflow: hidden;
 		z-index: var(--z-lifted);
-	}
+		width: fit-content;
 
-	.arrow {
-		position: relative;
-		top: 4px;
-		width: 20px;
-		height: 20px;
-		transform: rotate(45deg);
-		border-radius: 2px;
-		background-color: var(--clr-bg-1);
-		border: 1px solid var(--clr-border-2);
-	}
-	.info-tooltip {
-		background-color: var(--clr-bg-1);
-		border-radius: var(--radius-m);
-		padding: 4px 8px;
-		border: 1px solid var(--clr-border-2);
-		box-shadow: 0px 6px 30px 0px var(--fx-shadow-m);
+		&::before {
+			content: '';
+			position: relative;
+			top: 4px;
+			width: 20px;
+			height: 20px;
+			transform: rotate(45deg);
+			border-radius: 2px;
+			background-color: var(--clr-bg-1);
+			border: 1px solid var(--clr-border-2);
+		}
 	}
 </style>
