+++ packages/ui/src/lib/utils/transitions.ts
@@ -14,7 +14,7 @@ export function slideFade(node: Element, options: SlideParams): TransitionConfig
 }
 
 export function flyScale(
-	_: Element,
+	node: Element,
 	params: {
 		y?: number;
 		x?: number;
@@ -23,16 +23,17 @@ export function flyScale(
 		position?: 'top' | 'bottom';
 	} = {}
 ): TransitionConfig {
+	const nodeStyle = getComputedStyle(node);
+	const transformX = new WebKitCSSMatrix(nodeStyle.transform).m41;
+
 	// Default values
 	const DEFAULT_Y = -6;
-	const DEFAULT_X = 0;
 	const DEFAULT_SCALE_START = 0.94;
 	const DEFAULT_DURATION = 200;
 	const DEFAULT_POSITION = 'top';
 
 	// Extracting and using default values
 	const y = params.y ?? DEFAULT_Y;
-	const x = params.x ?? DEFAULT_X;
 	const startScale = params.start ?? DEFAULT_SCALE_START;
 	const duration = params.duration ?? DEFAULT_DURATION;
 	const position = params.position ?? DEFAULT_POSITION;
@@ -41,10 +42,9 @@ export function flyScale(
 		duration,
 		css: (t) => {
 			const translateY = y * (1 - t);
-			const translateX = x * (1 - t);
 			const scale = startScale + t * (1 - startScale);
 
-			return `transform: translate3d(${pxToRem(translateX)}, ${pxToRem(position === 'top' ? -translateY : translateY)}, 0) scale(${scale});
+			return `transform: translate3d(${transformX}px, ${pxToRem(position === 'top' ? -translateY : translateY)}, 0) scale(${scale});
 			        opacity: ${t};`;
 		},
 		easing: cubicOut
